Block unavailable times (#49)
* Remove unavailable times.

Closes #41.

* Block out used times.

This *mostly* solves the issue. You can still click a cell then hit space to select it, which kills the unavailability. I filter these out on save, but I'd still like to re-hide them if the user does this. I'll make a new issue for that.

* Add service account json to secrets.

* Fix typos in actions.

* Set up service account before package builds.

I also updated the token in SECRETS which was likely the issue. Hoping this takes care of it. It's also possible it's a single/double quote issue, so I changed that, too.

* Fix main GHA.

There was a hidden typo in this one so I copied from the working PR version.